refactor(mobile): remove native chat re-export facades
