+++ mobile/src/session/MobileNativeChatAsk.tsx
@@ -1,8 +1,8 @@
 import { useMemo, useRef, useState } from 'react'
 import { Pressable, ScrollView, StyleSheet, Text, TextInput, View } from 'react-native'
 import { Check } from 'lucide-react-native'
+import type { AskAnswerSelection, AskPrompt } from '../../../src/shared/native-chat-ask'
 import { colors, radii, spacing, typography } from '../theme/mobile-theme'
-import type { AskAnswerSelection, AskPrompt } from './mobile-native-chat-ask'
 
 type Props = {
   prompt: AskPrompt
