--- mobile/src/session/mobile-native-chat-ask.ts
@@ -1,2 +1,0 @@
-// Mobile and desktop share one parser/formatter so question-tool behavior cannot drift.
-export * from '../../../src/shared/native-chat-ask'
