--- mobile/src/session/mobile-native-chat-blocks.ts
@@ -1,11 +1,0 @@
-export {
-  isImageRefBlock,
-  isTextBlock,
-  NATIVE_CHAT_SOURCE_PRIORITY
-} from '../../../src/shared/native-chat-types'
-export {
-  foldToolMessages,
-  pairToolBlocks,
-  splitNativeChatBlocks,
-  type NativeChatToolPair as ToolPair
-} from '../../../src/shared/native-chat-tool-fold'
