+++ mobile/src/session/mobile-native-chat-controller-contract.ts
@@ -1,7 +1,11 @@
 import type { Dispatch, MutableRefObject, SetStateAction } from 'react'
+import type {
+  AskAnswerSelection,
+  AskPrompt,
+  parseAskFromStatus
+} from '../../../src/shared/native-chat-ask'
 import type { detectAgentPermission } from './mobile-native-chat-permission'
 import type { parseAgentQuestion } from './mobile-native-chat-question'
-import type { AskAnswerSelection, AskPrompt, parseAskFromStatus } from './mobile-native-chat-ask'
 import type { MobileNativeChatSendOutcome } from './mobile-native-chat-send'
 import type { MobileNativeChatPendingMessage } from './use-mobile-native-chat-drafts'
 import type { useMobileNativeChatSession } from './use-mobile-native-chat-session'
