--- mobile/src/session/mobile-native-chat-diff.ts
@@ -1,6 +1,0 @@
-export {
-  diffFromText,
-  diffFromToolCall,
-  type NativeChatDiffLine as DiffLine,
-  type NativeChatDiffLineKind as DiffLineKind
-} from '../../../src/shared/native-chat-diff'
