--- mobile/src/session/mobile-native-chat-merge.ts
@@ -1,9 +1,0 @@
-// Keep the established mobile import path while executing the shared desktop/mobile merger.
-export {
-  applyAppend,
-  boundNativeChatWindow,
-  createNativeChatMerger,
-  mergeNativeChatMessages,
-  replaceList,
-  type NativeChatMerger
-} from '../../../src/shared/native-chat-merge'
