--- mobile/src/session/mobile-native-chat-noise.ts
@@ -1,1 +1,0 @@
-export { isNoiseMessage, stripNoiseMessages } from '../../../src/shared/native-chat-noise'
