+++ mobile/src/session/mobile-native-chat-stream-frame.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
+import { createNativeChatMerger, replaceList } from '../../../src/shared/native-chat-merge'
 import type { NativeChatMessage } from '../../../src/shared/native-chat-types'
-import { createNativeChatMerger, replaceList } from './mobile-native-chat-merge'
 import { applyMobileNativeChatStreamFrame } from './mobile-native-chat-stream-frame'
 
 function message(id: string): NativeChatMessage {
