+++ mobile/src/session/mobile-native-chat-stream-frame.ts
@@ -1,5 +1,6 @@
+import { applyAppend, replaceList } from '../../../src/shared/native-chat-merge'
+import type { NativeChatMerger } from '../../../src/shared/native-chat-merge'
 import type { NativeChatMessage } from '../../../src/shared/native-chat-types'
-import { applyAppend, replaceList, type NativeChatMerger } from './mobile-native-chat-merge'
 
 export type MobileNativeChatStreamFrame = {
   type?: string
